Inject colours into viewer
