+++ sites/app/package.json
@@ -9,8 +9,8 @@
     "check": "tsc"
   },
   "dependencies": {
-    "@biongff/vizarr": "workspace:*",
-    "@biongff/anndata-zarr": "workspace:*",
+    "@biongff/vizarr": "workspace:viewer",
+    "@biongff/anndata-zarr": "workspace:anndata-zarr",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
     "@mui/icons-material": "^7.2.0",
