+++ viewer/src/index.tsx
@@ -5,6 +5,6 @@ export { default as Vizarr } from "./components/VizarrViewer";
 export type { VizarrViewerProps } from "./components/VizarrViewer";
 
 export { createViewer } from "./api";
-export type { VizarrViewer } from "./api";
+export type { VizarrViewer, labelColor } from "./api";
 
 export type { ViewState, ImageLayerConfig } from "./state";
